:sparkles: feat(interfaces): Implementation - Implemented interfaces tracker, aggregate, analytics, metadata
