+++ packages/lib/src/interfaces/dhis2/analytics.ts
@@ -1,0 +1,36 @@
+export interface AnalyticsHeader {
+    name: string;
+    valueType?: string;
+    type?: string;
+    hidden?: boolean;
+    meta?: boolean;
+}
+
+
+export interface AnalyticsItem {
+    [key: string]: {
+        name: string;
+        [key: string]: any;
+    }
+}
+
+export interface AnalyticsDimension {
+    dx?: string[];
+    pe?: string[];
+    ou?: string[];
+    co?: string[];
+}
+
+export interface AnalyticsMetadata {
+    items: AnalyticsItem[];
+    dimensions: AnalyticsDimension;
+}
+
+export interface Analytics {
+    headers?: AnalyticsHeader[];
+    metaData?: AnalyticsMetadata;
+    rows?: string[];
+    height?: number
+    headerWidth?: number
+    width?: number
+}
