+++ packages/lib/src/interfaces/dhis2/index.ts
@@ -1,0 +1,5 @@
+export * from "./base";
+export * from "./aggregate";
+export * from "./tracker";
+export * from "./metadata";
+export * from "./analytics";
